Tests/LibWeb/Text/input/display_list/contenteditable-empty-caret-padding.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    p {
        border: 3px solid green;
        margin: 0;
        min-height: 100px;
        padding: 24px;
        width: 300px;
    }
</style>
<p contenteditable></p>
<script>
    test(() => {
        document.querySelector("p").focus();
        const caretLine = internals.dumpDisplayList()
            .split("\n")
            .find(line => line.includes("FillRect"));
        println(caretLine);
    });
</script>
