Tests/LibWeb/Text/expected/display_list/contenteditable-empty-caret-padding.txt

  FillRect@1 rect=[35,35 1x18] color=rgb(0, 0, 0)
